+++ src/app/convenience-file-types.js
@@ -1,5 +1,23 @@
 import { FileType } from './constants';
 
+export function expandExpressionMatrixZarr(fileDef) {
+  const baseFileDef = {
+    ...fileDef,
+    coordinationValues: {
+      ...fileDef.coordinationValues,
+      obsType: fileDef.coordinationValues?.obsType || 'cell',
+      featureType: fileDef.coordinationValues?.featureType || 'gene',
+    },
+  };
+  delete baseFileDef.type;
+  return [
+    {
+      ...baseFileDef,
+      fileType: FileType.OBS_FEATURE_MATRIX_EXPRESSION_MATRIX_ZARR,
+    },
+  ];
+}
+
 export function expandRasterJson(fileDef) {
   const {
     url,
@@ -139,6 +157,7 @@ export function expandGenesJson(fileDef) {
  * expansion functions.
  */
 export const CONVENIENCE_FILE_TYPES = {
+  // [FileType.EXPRESSION_MATRIX_ZARR]: expandExpressionMatrixZarr,
   // [FileType.RASTER_JSON]: expandRasterJson,
   // [FileType.RASTER_OME_ZARR]: expandRasterOmeZarr,
   // [FileType.CELL_SETS_JSON]: expandCellSetsJson,